修改变量类型

+++ watcher-manager/src/main/java/com/aihuishou/bi/entity/User.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
 @Setter
 public class User implements Serializable {
 
-    private Long obId;
+    private Integer obId;
 
     private String name;
